% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{data}
\name{population}
\alias{population}
\title{Estimates of US population by year and sex}
\format{\describe{
 \item{year}{Range: 2000-2014}
 \item{age}{Range: 0-84}
 \item{sex}{male or female}
 \item{n}{Estimated population}
}}
\source{
The US census:
 \url{http://www.census.gov/popest/data/intercensal/national/nat2010.html}
 and \url{https://www.census.gov/popest/data/state/asrh/2014/SC-EST2014-AGESEX-CIV.html}
}
\usage{
population
}
\description{
Estimates of US population by year and sex
}
\keyword{datasets}

