% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\docType{data}
\name{products}
\alias{products}
\title{NEISS product codes}
\format{\describe{
   \item{code}{Product code}
   \item{title}{Description}
   \item{deleted}{Is the code still in active use?}
}}
\source{
\url{http://www.cpsc.gov/Media/Documents/Research--Statistics/NEISS-Injury-Data/Product-Code-Comparability-Table/} extracted with Tabula.
}
\usage{
products
}
\description{
Descriptions of the NEISS product codes.
}
\keyword{datasets}
