% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{data}
\name{population}
\alias{population}
\title{Estimates of US population by year and sex}
\format{\describe{
 \item{year}{Range: 2013-2016}
 \item{age}{Range: 0-84}
 \item{sex}{male or female}
 \item{n}{Estimated population}
}}
\source{
The US census:
 \url{https://wonder.cdc.gov/Bridged-Race-v2016.HTML}
}
\usage{
population
}
\description{
Estimates of US population by year and sex
}
\keyword{datasets}
